+++ package.json
@@ -59,6 +59,7 @@
     "@babel/register": "^7.6.2",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
+    "chai-http": "^4.3.0",
     "mocha": "^7.0.1",
     "nodemon": "^2.0.2",
     "sinon": "^8.1.1"
